[FEAT] Add PostgreSQL plugin and workflow repository APIs:
  Implemented PostgreSQL plugin for workflow persistence, including dual storage pattern for YAML and JSONB. Added `IWorkflowRevisionRepository` interface with methods for CRUD operations on workflows. Updated core logic for revision handling, integrating `WorkflowID`, `WorkflowRevisionID`, and `WorkflowRevisionWithSource`. Enhanced test coverage with `WorkflowIDTest`. Removed `ModelExceptions` and mapped exceptions to `MaestroException` hierarchy.

+++ api/src/main/kotlin/io/maestro/api/dto/WorkflowDtos.kt
@@ -10,7 +10,7 @@ import java.time.Instant
 data class WorkflowRevisionResponse(
     val namespace: String,
     val id: String,
-    val version: Long,
+    val version: Int,
     val name: String,
     val description: String,
     val active: Boolean,
